# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: cdCTF API
  version: 0.1.0
  description: cdCTF API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /learn/lessons:
    get:
      operationId: listLessons
      tags: [learn]
      summary: List lessons
      parameters:
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: Lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
  /learn/lessons/{id}:
    get:
      operationId: getLesson
      tags: [learn]
      summary: Get lesson detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /learn/lessons/{id}/test/start:
    post:
      operationId: startLessonTest
      tags: [learn]
      summary: Start a lesson test (activates anti-cheat)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestStartResponse"
        "403":
          description: Blocked or max attempts reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /learn/lessons/{id}/test/submit:
    post:
      operationId: submitLessonTest
      tags: [learn]
      summary: Submit lesson test answers
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitTestBody"
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResult"
        "403":
          description: Blocked or not started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /learn/lessons/{id}/test/escape:
    post:
      operationId: reportTestEscape
      tags: [learn]
      summary: Report ESC key press during test (anti-cheat)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Escape handled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscapeResponse"

  # SCOREBOARD
...
  /users/{id}:
    get:
      operationId: getUserProfile
      tags: [profile]
      summary: Get user profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateUserProfile
      tags: [profile]
      summary: Update user profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/competitions/{id}/users/{userId}:
    post:
      operationId: adminAddCompetitionUser
      tags: [admin]
      summary: Add user to private competition
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth
...
    Lesson:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        titleUz:
          type: string
          nullable: true
        titleRu:
          type: string
          nullable: true
        categoryId:
          type: integer
        categoryName:
          type: string
        points:
          type: integer
        isCompleted:
          type: boolean
        isBlocked:
          type: boolean
        attemptCount:
          type: integer
      required: [id, title, categoryId, categoryName, points, isCompleted, isBlocked, attemptCount]
    LessonDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        titleUz:
          type: string
          nullable: true
        titleRu:
          type: string
          nullable: true
        content:
          type: string
        contentUz:
          type: string
          nullable: true
        contentRu:
          type: string
          nullable: true
        categoryId:
          type: integer
        categoryName:
          type: string
        points:
          type: integer
        isCompleted:
          type: boolean
        isBlocked:
          type: boolean
        attemptCount:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/LessonQuestion"
      required: [id, title, content, categoryId, categoryName, points, isCompleted, isBlocked, attemptCount, questions]
    LessonQuestion:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        questionUz:
          type: string
          nullable: true
        questionRu:
          type: string
          nullable: true
        options:
          type: array
          items:
            type: string
        optionsUz:
          type: array
          items:
            type: string
          nullable: true
        optionsRu:
          type: array
          items:
            type: string
          nullable: true
      required: [id, question, options]
    TestStartResponse:
      type: object
      properties:
        sessionId:
          type: string
        attemptsLeft:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/LessonQuestion"
      required: [sessionId, attemptsLeft, questions]
    SubmitTestBody:
      type: object
      properties:
        sessionId:
          type: string
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: integer
              selectedOption:
                type: integer
            required: [questionId, selectedOption]
      required: [sessionId, answers]
    TestResult:
      type: object
      properties:
        passed:
          type: boolean
        score:
          type: number
        correctCount:
          type: integer
        totalCount:
          type: integer
        attemptsLeft:
          type: integer
        pointsEarned:
          type: integer
      required: [passed, score, correctCount, totalCount, attemptsLeft, pointsEarned]
    EscapeResponse:
      type: object
      properties:
        escapeCount:
          type: integer
        blocked:
          type: boolean
        timeoutSeconds:
          type: integer
          nullable: true
        message:
          type: string
      required: [escapeCount, blocked, message]

    # Scoreboard
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        nickname:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
          nullable: true
        points:
          type: integer
        rank:
          type: integer
        isBlocked:
          type: boolean
        solvedCtf:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              category:
                type: string
              points:
                type: integer
              solvedAt:
                type: string
                format: date-time
            required: [id, name, category, points, solvedAt]
        completedLessons:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              title:
                type: string
              points:
                type: integer
              completedAt:
                type: string
                format: date-time
            required: [id, title, points, completedAt]
        competitionHistory:
          type: array
          items:
            type: object
            properties:
              competitionId:
                type: integer
              competitionName:
                type: string
              rank:
                type: integer
              points:
                type: integer
            required: [competitionId, competitionName, rank, points]
        titles:
          type: array
          items:
            $ref: "#/components/schemas/Title"
        createdAt:
          type: string
          format: date-time
      required: [id, nickname, email, points, rank, isBlocked, solvedCtf, completedLessons, competitionHistory, titles, createdAt]
    UpdateProfileBody:
      type: object
      properties:
        nickname:
          type: string
          minLength: 3
          maxLength: 32
...
    Title:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        points:
          type: integer
          description: Bonus points for earning this title
        earnedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, name, category, points]

    # Admin
...